Dritte Übung

+++ app/src/main/java/at/fh/mappdev/loggingviewsandactivity/LessonRating.kt
@@ -1,4 +1,4 @@
 package at.fh.mappdev.loggingviewsandactivity
 
-class LessonRating( val ratingValue: Int, val feedback: String,  ) {
+class LessonRating( val ratingValue: Double, val feedback: String,  ) {
 }
